have sort functionality working on first MCSP table. Need styling and dynamic functionality to useSortableData as it only pulls the 1st currCohort students. ETS dates don't filter due to formatting need to troubleshoot solution.
